+++ R/qualpal.R
@@ -66,6 +66,8 @@
 #'  close to the background/canvas color. If \code{NULL} (the default), the
 #'  background color is not considered at all. Any color that is convertable
 #'  via \code{\link{col2rgb}} is acceptable, including hex colors.
+#'  @param metric The color metric to use for the color distance
+#'    matrix.
 #'
 #' @return A list of class \code{qualpal} with the following
 #'   components.
@@ -111,7 +113,8 @@ qualpal <- function(
   colorspace = list(h = c(0, 360), s = c(0.2, 0.5), l = c(0.6, 0.85)),
   cvd = c("protan", "deutan", "tritan"),
   cvd_severity = 0,
-  bg = NULL
+  bg = NULL,
+  metric = c("din99d", "ciede2000", "cie76")
 ) {
   UseMethod("qualpal", colorspace)
 }
@@ -278,9 +281,11 @@ predefined_colorspaces <- function(colorspace) {
 make_options <- function(
   cvd = c("protan", "deutan", "tritan"),
   cvd_severity = 0,
-  bg = NULL
+  bg = NULL,
+  metric = c("din99d", "ciede2000", "cie76")
 ) {
   cvd <- match.arg(cvd)
+  metric <- match.arg(metric)
 
   stopifnot(
     is.character(cvd),
@@ -301,7 +306,7 @@ make_options <- function(
   } else {
     bg <- double(0)
   }
-  list(cvd = cvd_list, bg = bg)
+  list(cvd = cvd_list, bg = bg, metric = metric)
 }
 
 validate_args <- function(n) {