+++ tests/testthat/test_qualpal.R
@@ -26,6 +26,7 @@ test_that("erroneous input to qualpal() returns errors", {
   expect_error(qualpal(3, matrix(1:9, ncol = 3)))
   expect_error(qualpal(3, matrix(runif(10), ncol = 5)))
   expect_error(qualpal(3, bg = "notacolor"))
+  expect_error(qualpal(3, metric = "HSL"))
 })
 
 test_that("proper use of qualpal() works", {
@@ -43,4 +44,5 @@ test_that("proper use of qualpal() works", {
   ))
   expect_silent(qualpal(3, bg = "white"))
   expect_silent(qualpal(9, bg = "#00ff00"))
+  expect_silent(qualpal(9, metric = "ciede2000"))
 })